[libromdata] Xbox360_XEX: Unreference the RomData subclass on error, not the file.
The file will be unreferenced anyway, so unreferencing it twice results
in an assertion.

The RomData subclass maintains its own reference to the file and unref()'s
it when it's deleted. We get our own reference when we open the file, so
we should only unref() it once.

This fixes cases in which the EXE and/or XDBF can't be opened for some
reason, e.g. when testing XEX1 executables. (XDBF seems to have a
different format for XEX1...)