[gtk] Split rom-properties.thumbnailer into a separate package, rom-properties-gtk3-common.
Moved it up from the gnome directory to the gtk3 directory.

NOTE: We can't add a MATE package right now because the earliest version
of Ubuntu I'm still supporting is 16.04, and that has MATE 1.12 (GTK2).
I'll add the MATE (and XFCE GTK3) packaging files in branches for the
appropriate Ubuntu verisons:

- Ubuntu 16.04 LTS: MATE 1.12.7 (GTK2), Thunar 1.6.11 (GTK2)
- Ubuntu 18.04 LTS: MATE 1.20.2 (GTK3), Thunar 1.6.15 (GTK2)
- Ubuntu 18.10: MATE 1.20.2 (GTK3), Thunar 1.8.1 (GTK3)
- Ubuntu 19.04: MATE 1.20.3 (GTK3), Thunar 1.8.4 (GTK3)