Removed S2TC decoding and test images.
S2TC was originally implemented because the S3TC patents were still in
effect. The patents expired in October 2017, which was *before* the
release of v1.2 (the first version with S3TC decoding), but I didn't want
to remove it so close to release.

Pretty much no one used the S2TC version, so it was just wasting space.

Code size reduction: (64-bit Gentoo Linux, gcc-9.1.0, release build)

   text    data     bss     dec     hex filename
  17058       9       0   17067    42ab ImageDecoder_S3TC.cpp.o [before]
  13138       8       0   13146    335a ImageDecoder_S3TC.cpp.o [after]
  -3920      -1       0   -3921    -f51 Difference