[libromdata] SegaPVR: Added format detection for SVR textures.
It's technically a "PVR" texture, but the image data types and pixel
formats are different, so it's handled as a special case of PVR.

systemName(): Handle SVR differently.

TODO: Some of the sample SVR textures I'm using have an image data type
of 0x61, which isn't defined in Puyo Tools.