[libromdata] SegaPVR: The 32-bit SVR format is actually BGR888_ABGR7888.
If the high bit is set, then the value is interpreted as BGR888.
Otherwise, it's interpreted as ABGR7888.

This is a really stupid system and is unnecessary, since you can just
set the alpha channel to 0xFF. It doesn't save any space.

This makes the textures look closer to how they should appear, though
there's still issues with the alpha channel not being quite right.