[libromdata] SegaPVR: Format 0x61 appears to be Rectangle, but sometimes swizzled.
If the pixel format is BGR5A3, and the texture size is >=64x64, then the
texture is swizzled.

FIXME: The 4-bit/8-bit unswizzle function is *NOT* correct for 16-bit.
It's being used for now as a placeholder.

FIXME 2: Puyo Tools doesn't support format 0x61.

[librpbase] ImageDecoder::fromLinear16(): Handle BGR5A3.
ImageDecoder::fromLinear16_sse2(): Ignore BGR5A3.