[cmake] Use the LDFLAG -Wl,--compress-debug-sections=zlib instead of specifying --compress-debug-sections in objcopy.
FreeBSD's objcopy doesn't have --compress-debug-sections, but ld does.

On Linux, this does provide an advantage over the previous method because
ld compresses the debug sections when creating the original executable,
which results in less disk access.

Three compression options are available in GNU ld:
- zlib-gnu: Use .zdebug_* sections.
- zlib-gabi: Use .debug_* sections with the C flag.
- zlib: Same as zlib-gabi.

FreeBSD's zlib option is the same as GNU ld's zlib/zlib-gabi option,
so we'll use it on all platforms.

NOTE: FreeBSD's default toolchain is llvm/clang as of FreeBSD 10.0, so
this should apply to other systems using llvm/clang.