[librpbase] Implemented SCSI support for FreeBSD, but it isn't quite working.
It detects the full size of the Xbox 360 disc, but attempting to read
past the DVD-ROM area results in a SCSI ILLEGAL REQUEST error 21,0:
Logical block address out of range.

Reworked the SCSI implementation to share more stuff between Linux and
Windows. This includes splitting OS API disk space code to a new function,
rereadDeviceSizeOS(). This has the code that was previously in the
RpFile_win32.cpp file for Windows, and new code for Linux.

The Linux version also uses readUsingBlocks() now.

Support on both Linux and Windows is basically the same as it was before.