[librptexture] FileFormat::getFields(): New function to get texture-specific fields.
This is conditionally compiled-in by setting ENABLE_LIBRPBASE_ROMFIELDS.
It's always enabled at the moment. The main purpose is to allow the
library to be used by other programs at some point in the future,
while retaining full functionality for rom-properties.

ValveVTF: Fixed some inconsistencies with flags due to some flags being
nullptr while enumerating them backwards. The flags are now enumerated
forwards, which might be slightly less efficient, but it ensures that
the flags are correct.