[gtk] RomDataView: Make sure the ListData icon is not nullptr before attempting to use it.
Currently having an issue with "You Don't Know Jack" (Xbox 360) where the
Xbox360_XDBF class isn't decoding two of the icons, resulting in nullptr.
This shouldn't cause a problem in the release build, since the
GdkImageConv() function will simply ignore it, but in debug builds, it
loudly assert()s.

Also: Millionaire 2012 (Xbox 360) fails with an assertion somewhere else
in the avatar awards code.