[libromdata] Xbox360_XDBF.cpp: Don't increment icon_iter if the icon is being skipped.
This fixes the assertion for YDKJ, though it's skipping two avatar awards.
abgx360 shows them as being different awards, even though they have the
same ID number. The awards are the same except for gender, so I'm guessing
it has something to do with the way the avatar system works.

We should remove the duplicate avatar ID check and show all of them.