[librpbase] RomMetaData: Handle PropertyType::String with a nullptr value.
We'll just add a nullptr string.

TODO: Skip the property entirely? An empty string isn't expected here,
and that usually indicates something's wrong.

Added an assert() for debug builds.

[kde, win32] Handle nullptr strings by not adding the property.

This fixes issue #195: certain games crash explorer when moused over or
selected. Well, at least it fixes the crash; the underlying issue is in
the GameCubeBNR parser, and that needs to be fixed.