[libromdata] GameCubeBNR::loadMetaData(): Fix gamename_full vs. gamename error.
This was the underlying cause of the crash in issue #195.

Prince of Persia: The Sands of Time has an empty "full" game name. We
checked that first, saw it was empty, then went to the regular game name.
That one isn't empty, but due to a copy/paste error, the full game name
was still used, resulting in an empty entry.