[libromdata] EXE_PE: Added runtime DLL checks.
This checks for an MSVC runtime DLL entry in the import table.
If found, it reports the version and if it's a debug build.

New function pe_vaddr_to_paddr() that converts a virtual address to a
physical address. This only works within a single section, so a size
parameter is required to specify the desired range. If the range isn't
fully within a single section, or the virtual address isn't found, then
the function will return 0.

TODO:
- Return a download link if available.
- Add: MinGW, Cygwin, Visual Basic, 16-bit...