[kde] Disable deprecated Qt functionality in debug builds.
Not in release builds; this prevents breakage when a new Qt version is
released. Then again, Qt's implementation requires specifying a "last
tested" version, which also helps to prevent breakage...

ImageTypesTab: Use a lambda function instead of QSignalMapper for Qt 5.
QSignalMapper was marked as deprecated in Qt 5.10. Note that the lambda
function works for all releases of Qt 5, so we're using it for all
Qt 5.0 and later. (The QSignalMapper code is retained for Qt 4.)