[svrplus] add message loop
Right now this works similarly to the DialogBox function.

Dialog end is done differently. EndDialog sets a flag in a struct and
sends a WM_NULL to the window. The message loop checks the flag on each
iteration, and if it's present, it terminates the loop, saves the return
value (also from the same struct), and calls DestroyWindow.

Aforementioned struct (called DIALOGINFO in Wine/ROS) is undocumented,
so we have to do this ourselves. For simplicity, EndDialog is replaced
with DestroyWindow+PostQuitMessage. This is less robust, but good enough
for our purposes.