[libromdata] Reorder ROM detection to fix detection of a particular Nintendo DS ROM image.
The Nintendo DS ROM image "Live On Card Live-R DS" was misdetected as
an Xbox 360 STFS package because the first four characters of the ROM
image is the word "LIVE", which is the magic number for one type of
STFS package.

Moved Xbox360_STFS to the end of the list.

Added magic numbers for Nintendo DS so we can detect it earlier.

This bug was reported by @Whovian9369.