[librpbase/tests] TextFuncsTest: Test Atari ST and ATASCII conversion functions.
These aren't currently used, but they might be used later.

TODO: Test PETSCII, and maybe add other variants of PETSCII.

str8_to_utf8(): Use a custom trailing NULL check. ATASCII uses \x00 as
a graphics character, so we need to allow it in the string. We'll at
least allow it as a character in the middle of the string; if it's at
the end of the string, it will be truncated.

Added ATASCII code points 0x02 and 0x0D.
Removed ATASCII code point 0xFC because it was incorrect.

This increases code coverage of TextFuncs_conv.cpp from 0% to 82.4%
for lines, and 0% to 75.0% for functions.

Total code coverage change:

Before:
  lines......: 15.7% (5090 of 32502 lines)
  functions..: 14.2% (335 of 2354 functions)

After:
  lines......: 15.7% (5118 of 32511 lines)
  functions..: 14.4% (338 of 2354 functions)