[librptexture] Implemented initial 4bpp decoding.
I'm not sur if RGB565 is correct. Sonic/roll005.tex looks like it has
the right blue, but there's weird artifacting. This corresponds with
the original data, so maybe it's some weird thing they did to compensate
for a low-quality screen. Regardless, those should probably be blue
pixels, not green.