[librptexture] DidjTex: Format 1 is linear RGB565.
The images using RGB565 still have the weird dithering, but it's
definitely less noticeable than the other images, so those are definitely
not in RGB565 format.

Renamed the width variables. The pow2 sizes are the "actual" sizes, and
the other sizes are the "display" sizes. The texture is loaded using the
"actual" size and displayed using the "display" size.

rom-properties uses the "actual" size, since it shows the raw texture.