[win32] RP_ThumbnailProvider: Check that file is not NULL before calling unref().
If the file is locked, e.g. NTUSER.DAT, then it won't be opened,
so d->file will be nullptr. However, we don't check for this, so
we end up calling unref() on nullptr, which crashes.

On Windows Vista and later, this only crashes the COM Surrogate,
not Windows Explorer. (if registered in release mode, that is)

Fixes #207: Win32 thumbnailer crashes on locked files

This bug was reported by @DankRank.

NOTE: This is a candidate for rom-properties 1.4.4.