[libromdata/tests] Added PowerVR3 tests.
FIXME: Several of these tests are failing. This might be due to a bug in
PVRTexToolCLI, which is what I used to create the expected images, but
more likely it's a bug in rom-properties. Leaving these tests disabled
for now.

Formats affected:
- PVRTC-I 4bpp RGB
- PVRTC-I 4bpp RGBA
- LA88
- RGBA8888

TODO: Add RGB9_E5 test images after removing the unneeded cube faces.