[libromdata] NCCHReader: Check for "icon" if it's a CFA.
DLC CIAs are considered CFAs, so the NCCH doesn't have a ".code"
section. The first file is "icon", so check for that instead.

This fixes icon display for DLC CIAs.

Reported by CheatFreak in issue #208:
No icon on some 3DS games even with all keys present and valid.

TODO: Check if this fixes the original issue in #208.