[libromdata] EXE_PE.cpp: Handle the import directory table in cases where it isn't a multiple of sizeof(IMAGE_IMPORT_DIRECTORY).
We need to keep the weird RVA code because MSVC 2017 stores the
filenames before the table, at least in some cases.

Added an le32_to_cpu() for rvaModuleName.

NOTE: The le32_to_cpu() most likely still applies for Xbox 360, even
though it's big-endian, since everything else in the PE executable is
left as-is. However, the import directory table is probably not used in
Xbox 360 executables, so it isn't important. Never mind the fact that
Xbox 360 executables wouldn't be using one of these runtimes, anyway.