[librpbase] Implemented mtime functionality in the Win32 version of RpFile::setOriginInfo().
This uses SetFileTime() instead of the CRT function _futime64().

[libcachemgr] UrlmonDownloader: GetUrlCacheEntryInfo() might fail with
last error == ERROR_INSUFFICIENT_BUFFER, which is a valid result.
I'm not sure when this behavior changed, or if it was always like this...

FIXME: amiibo.life files aren't being found in the cache. This might be
because the files are redirected to a CDN, and we need to use the CDN
URL instead of the original URL.