[gtk] CreateThumbnail.cpp: Added support for URIs.
The stub function takes either a filename or a URI, and it's auto-detected.

Note that this currently only works on GTK+ desktops. On KDE, the KDE
plugin will be selected, and it doesn't support non-local files yet
(and probably not even file:// URIs).

rom-properties.thumbnailer.in: Pass a URI instead of a filename.