[libcachemgr] Initial rp-download implementation.
rp-download currently takes a cache key as a parameter and prints both
the source URL and the destination filename in the local cache.

rp-download has its own logic for determining the URLs based on the
cache key prefix (the first component). Any unsupported prefix will
be ignored.

[libcachecommon] CacheKeys: Added wchar_t versions to reduce character
set conversion in rp-download.