[rp-download] rp-download.cpp: Fix prefix length for Wii U.
'5' included the slash, which was originally going to be part of the
check, but then I added the separate prefix_len check, so we don't have
to check for the slash. Without prefix_len, we'd have to check for the
slash in order to distinguish between e.g. "wii" and "wiiu".