[kde] RomDataView: Initial support for RFT_MULTI_STRING.
initStringMulti(): Create the labels, but don't actually fill them in.
Save the QLabel* and the RomFields::Field* in a QVector for later.

updateStringMulti(): Update *all* RFT_MULTI_STRING labels all at once.
This will be called after initialization and if the user changes the
selected language.

initDisplayWidgets(): Call updateStringMulti() after initializing.

TODO: Add a dropdown box, defaulting to def_lc, and allow the user
to select a different language.