[win32] RP_ShellPropSheetExt: Added a combobox for language selection.
Not functional yet.

s/hTabWidget/tabWidget/

Store the header row's RECT in the private class so we can use it
when creating cboLanguage.

TODO: We might need to create cboLanguage when creating the header row
in order to preserve tab order. If we do this, it'll be hidden initially.
Need to check all three frontends.

TODO: Convert lc to string if lc_str is invalid.

[libwin32common] WinUI.cpp: s/errro/errno/