[libromdata] WiiPartition: Calculate the used partition size correctly for unencrypted RVT-H Reader disc images.
Previously, GcnPartition was always multiplying by 32/31, regardless of
the encryption status.

The Last Story (PAL) v2340 has ~4.70 GiB of data when extracted, including
the FST and main.dol. rpcli was previously showing 4.85 GiB used; now it
shows 4.70 GiB, which includes the FST, main.dol, and partition header.

Also check if the DOL is after the FST, and if it is, use the DOL offset.
TODO: Get the DOL size if the DOL is located after the FST. This is
unlikely, though...