[libromdata] Xbox360_XDBF: Find all the string table indexes on startup.
findResource() is O(n). The new functin initStrTblIndexes() does a single
scan of the entry table to find all of the string tables instead of
scanning it multiple times. The indexes (not pointers) are then stored
in an std::array.

loadStringTable(): Use strTblIndexes for the string table indexes.

This will make it easier to implement multiple language support.