[librpbase] RomFields: Added cbegin() and cend() for RomFields::Field access.
Renamed field() to at() for consistency with STL.

Updated everything to use cbegin() and cend(), with a cached cend() due
to the pimpl pattern. (Note that LTO would probably optimize this out.)

Note that this does expose the internal type of RomFields due to the
direct use of std::vector<RomFields::Field>::const_iterator.

[win32] RP_ShellPropSheetExt: Even though we're using iterators now, we
still need the original index value for control IDs.

TODO:
- UI frontends: Pass RomFields::Field objects by const ref instead of
  const pointer so we can get rid of nullptr checks.