[librptexture] S3TC, BC7: Allow image sizes that aren't a multiple of the tile size.
DDS in particular allows this. The end result is the tile size is rounded
up, and the right-most columns and/or bottom-most rows are truncated.

Also added support for KTX and VTF.

NOTE: GameCube DXT1 does *not* allow this and will still assert().

TODO: Actually truncate the size. Currently, the returned image will have
the last columns/rows intact.

This bug was reported by @HyperPolygon64.