[librptexture] rp_image, rp_image_backend: Added a shrink() function for S3TC/BC7.
rp_image_backend_default: shrink() adjusts width and height in-place.
The row stride ensures that the image is read correctly. This makes it
so we don't have do an image data copy.

RpQImageBackend, RpGdiplusBackend: TODO: Figure out how to resize the
image without copying. For GDI+, we can probably do some fiddling with
Lock(), though that might cause problems when converting to HBITMAP.

ImageDecoder S3TC/BC7: Shrink the image if it isn't a multiple of the
tile size.