[librpbase] RomFields::strArrayToVector*(): Removed the NULL-terminated array handling.
It didn't work, since we checked strArray for NULL, not *strArray.

NULL pointers in the array is handled as an empty string, and some
things might make use of this.

All callers do specify an explicit count, so remove the -1 handling and
change count to size_t.

NOTE: For vector<string>::emplace_back(), specifying no arguments is
equivalent to calling the no-argument std::string constructor.
Specifying string() creates a new object and then copies it. Specifying
"" is equivalent to the single-argument std::string constructor.

MSVC 2010 doesn't support the no-argument version, so we'll use "".

TODO: Switch to the no-argument version when we no longer support
MSVC 2010. (Or, add conditional compilation...)