[win32] RP_ShellPropSheetExt: Reduced the scope of hdcIcon to prevent a leak.
Previously, it would skip over the icon code if the icons couldn't be
loaded, and then release the DC. When the function to create the ImageList
was split out of the main RFT_STRING_MULTI code, it returned immediately
if the icons couldn't be loaded, and forgot to release the DC.

This fixes a regression from commit da29cf876fef3635e8b685f8b3aa30e68d976fa8.
([win32] RP_ShellPropSheetExt: Split building the cboLanguage ImageList into a separate function.)

PVS-Studio V773 (High): The function was exited without releasing the
'hdcIcon' handle. A resource leak is possible.