RomThumbCreator, CreateThumbnail: Inline some small functions into the class definition.
I stopped doing this a while back because Mac OS X with gcc-4.2 would
miscompile these functions (incorrect `this` pointer). I'm not targetting
Mac OS X yet here, and modern Mac OS X uses LLVM/clang, not gcc, so this
shouldn't be a problem anymore.

This does affect the code size slightly. On 64-bit Linux in a release
non-LTO build, the GTK+ .so files' .text sections increase by 48 bytes,
and the KF5 .so file .text section increases by 128 bytes. In an LTO
build, it's +93 bytes for GTK+, +172 bytes for KF5.