[cmake] PrecompiledHeader.cmake: Use ${CMAKE_${_lang}_FLAGS}.
This should ensure better utilization of precompiled headers in gcc builds.

Note that ${CMAKE_BUILD_TYPE} is not necessarily all uppercase, so
convert it to uppercase before using it.

CMakeLists.txt: Make sure CMAKE_BUILD_TYPE gets cached as "Release" if it
isn't specified on the command line.

TODO: Do we need to ensure that ADD_COMPILED_HEADER() is added *after* all
flags/definitions are set, or does the ordering not matter?