[win32] RP_ShellPropSheetExt: Delete the bitmap after adding it to the ImageList.
ImageList makes a copy of the bitmap, so we have to delete it when
we're done with it. Otherwise, we end up leaking the bitmap.

This fixes a regression from commit da29cf876fef3635e8b685f8b3aa30e68d976fa8.
([win32] RP_ShellPropSheetExt: Split building the cboLanguage ImageList into a separate function.)