[win32] RP_ExtractIcon: Make sure hbmpImage is deleted in the error path.
The error path isn't likely to be taken, but just in case...

Also, use DestroyIcon() and DeleteBitmap() to delete icons and bitmaps,
not DeleteObject(). DeleteBitmap() is a macro that calls DeleteObject()
internally, but DestroyIcon() does other stuff. This might hae resulted
in a memory leak if an error path was taken, but again, this error path
(a different one) was unlikely to be taken.