[gtk] CreateThumbnail.cpp: Simplified filename handling.
New function openFromFilenameOrURI() that opens the source file and
determines if it's a filename or a URI. It returns the normalized URI
using an std::string reference.

Note that this does have extra overhead in the case where the source
file is a URI, since previously the string wouldn't be copied, but it
reduces cognitive overhead by not having to deal with special cases.

openFromFilenameOrURI() returns the file using an IRpFile**.

The return value is either 0 on success or an RPCT_* value on error.