[libromdata] XboxDisc: Simplify the XGD3 check.
The previous method was incorrect, since it was adding 1 to the
`const char*` pointer when it really needed to add 10. Coverity
complained about a potential buffer overflow because of it.

Simplified by doing two explicit memcmp() calls instead of using a loop.

This bug has been present since XboxDisc was introduced in
commit 4584be137faffb8e8cb75b661ff1cbfd108db7ec.
([libromdata] XboxDisc: Initial parser for XDVDFS disc images.)

NOTE: This bug is present in v1.4, but it likely won't cause any issues
in practice (unless someone is intentionally fuzzing a disc image).

This fixes the following Coverity Scan bugs:
- #1483989: Out-of-bounds access