[kde] RomDataView::eventFilter(): Removed a line of debugging code.
This prevented RFT_LISTDATA views from being resized properly for e.g.
Nintendo3DS.

This fixes a regression from commit 9a6ad950e5554135cf505d2f65a94b34b27e616c.
([kde] RomDataView: If the last field in a tab is RFT_LISTDATA in a separate row, move it out of the QFormLayout and into the QVBoxLayout.)

NOTE: This bug is present in v1.4. It's a minor display issue, but it can
be annoying if it results in the RFT_LISTDATA views being too small.

This fixes the following Coverity Scan bugs:
- #1483974: Structurally dead code