Windows: Use FindFirstFile() instead of stat() or other functions.
stat() is implemented using either FindFirstFile() or
GetFileAttributesEx() depending on MSVCRT version. There's also extra
processing needed, since Windows doesn't usually maintain all of the
attributes found in `struct stat`, so MSVCRT has to fill them in
manually.

Use FindFirstFile() directly, since we only need the file size, time,
and possibly attributes.

Note that we have to use FileTimeToUnixTime() to convert the Win32
FILETIME to time_t. MSVCRT's stat() does this for all three timestamps,
so this is a net win for us because we only have to do this on the
mtime field.

FileSystem_win32.cpp: Previously, we opened the file and then used
GetFileSizeEx() and GetFileTime(). FindFirstFile() works better here
because it gets both size and time at the same time, plus we can get
the file attributes to determine if it's a directory, and we don't
have to open the file manually.