[libromdata] Xbox360_XEX: Use reader.size() instead of ARRAY_SIZE(reader).
It's an std::array<>, so use the STL-like function.

PVS-Studio V651 (High): An odd operation of the 'sizeof(X)/sizeof(X[0])'
kind is performed, where 'X' is of the 'class' type. consider inspecting
the 'sizeof (reader) / sizeof (reader[0])' expression.

FIXME: Same warning in NintendoBadge.cpp for a static assertion, but we
can't use std::array<>::size() because it isn't constexpr. (Well, the
documentation says it is, but MSVC 2019 says it isn't.)