[librpbase/tests] gtest_init.cpp: Use _tputenv() instead of the various versions.
We're only setting the LC_ALL environment variable, so we don't need to
use setenv() or _wputenv_s().

[librpbase] Removed the putenv()/setenv()/etc. checks.

tcharx.h: Added a _tputenv() wrapper macro.

NOTE: On POSIX systems, putenv() makes the passed-in `char*` part of the
environment. Hence, we have to keep it in a `static char[]` (not const),
since it's technically editable.