Fix `extern TCHAR dll_filename[]` on MinGW-w64.
MSVC exported it from DllMain.cpp with C++ name mangling, and imported
it in ExecRpDownload_win32.cpp with C++ name mangling as well.

MinGW-w64 exported it with C name mangling, not C++. Hence, we had to
add `extern "C"` to ExecRpDownload_win32.cpp.

In order to ensure that this still works on MSVC, it has to be exported
using `extern "C"` in DllMain.cpp as well.

RP_ShellIconOverlayIdentifier.cpp: Added `extern "C"`; use TCHAR instead
of wchar_t.