[librptexture] rp_image::squared(): Added (non-optimal) CI8 support; fix issues with images that are taller than they are wide.
The sq_img was always created using width as the dimension, so if the
original image is taller, this would result in memory corruption.

Factored out some common code for getting the source and destination
buffers and stride values.

This showed up when viewing some Didj textures on Windows. This only
affects Windows, since rp_image::squared() is currently only used for
Windows icons. (...though it might be used for thumbnails too?)