[libromdata] DMG: Added external title screen images using my new RPDB subdomain.
Example URL: https://rpdb.gerbilsoft.com/gb/title/E/BYTE01.png

For CGB games that have a valid game ID, the game ID is used for the
filename, and the region byte is used as the subdirectory.

To improve game ID detection, getTitleAndGameID() now restricts valid
game IDs to [ABV] and region codes to [ADEFGHIJKPSXY]. This fixes
false positives with some games that have e.g. 'CKET' at the end of
the title.

For games without a game ID, the filename consists of the title and the
publisher code. If the region flag is Japanese, a "-J" is appended.

For certain DMG/CGB titles, the global checksum from the ROM header is
also appended:
- Pokémon Gen1 games (non-CGB, non-JP only)
- The Legend of Zelda: Link's Awakening DX (CGB, non-JP only, except
  US v1.2)

TODO:
- Remove leading and trailing spaces from titles?
- There's at least 100 or so ROM images from GoodGBx that aren't
  thumbnailed yet because they need some more special cases.