[libromdata] Xbox360_XEX: Initial support for XEX patches.
XEX patches have the same structure as XEX executables, but if they have
the Delta Patch flag set, the contained executable isn't a standard PE
executable; it's a patch.

We'll indicate the file is a patch if any of the patch flags are set, and
skip over EXE and XDBF parsing if Delta Patch is set.

[librpbase] RomData: New FTYPE_PATCH_FILE file type.

[xdg] rom-properties.xml: Detect XEX patches. This checks all three flags,
not just the Delta Patch flag.