[libromdata] Xbox360_STFS: Attempt to open default.xex or default.xexp.
Parse the file table and, if one of these is found, open it using the
Xbox360_XEX class.

FIXME: Hash tables aren't handled properly, so only the file header
can be reliably read properly. This works for default.xexp, since we
can't handle the delta patches anyway, but will result in things like
the achievements table not showing up for default.xex.