[libromdata] WiiWAD: Fix broken DLC icons caused by CBCReader fixes.
CBCReader was updated to properly update its current position, which was
needed for Xbox 360 XDBF files.

WiiWAD incorrectly depended on it *not* updating, since we were checking
for a content.bin header, which doesn't exist in WAD files. Don't check
for the content.bin header anymore.

This fixes a regression from commit e2553d5bc14e527b5483cb3431f2fc348fc6f236.
([librpbase] CBCReader: Make sure d->pos stays in sync with the read position.)

NOTE: This means this feature is broken in v1.4 (when it was introduced)
and in v1.5.