[librpbase] Config: Added DMG config options to select title screen images to use for different Game Boy models.
These options in the [DMGTitleScreenMode] block can be used to set which
image is used for DMG (Game Boy), SGB (Super Game Boy), and CGB (Game Boy
Color) ROM images.

- If a ROM supports CGB mode, the CGB value is used.
- If a ROM supports SGB mode, the SGB value is used.
- Otherwise, the DMG value is used.

These values can be set to DMG, SGB, or CGB, which represents how the
particular ROM shows up on that particular model.

[libromdata] DMG: Use the DMG config options to select a subdirectory
on the RPDB server.