[win32] RP_ExtractImage: Return E_FAIL if the file could not be opened.
Otherwise, we attempt to use the unref()'d file, which crashes.

This crash is reproducible on Windows if a shortcut file (.lnk) is
pointing to a file type supported by rom-properties, but the target
file is missing. (dangling shortcut)

This fixes issue #217: Windows 7 explorer crashes constantly after
installing rom propreties (reported by @Midou36O).

This commit is a candidate for rom-properties 1.5.1.