Initial implementation of "achievements".
Currently, three achievements are defined:
- Viewing a debug-encrypted file.
- Viewing a non-x86/x64 PE executable.
- Viewing a BroadOn WAD file.

The only libromdata subclass that implements this functionality right
now is WiiWAD for debug encryption and BroadOn WAD files.

AchLibNotify: Implementation that calls `notify-send`. This should be
replaced with D-Bus access, since the `notify-send` executable takes
time to run, which results in the caller hanging for a bit.

Added achievement checks to the GTK+, KDE, and Win32 frontends. Note
that the Win32 version doesn't actually work yet, since we don't have
a Win32 implementation right now.

TODO: Move the implementation out of librpbase and into the UI frontends.