[libromdata] GameCube: Added the ViewedDebugCryptedFile achievement.
GameCubePrivate::loadWiiPartitionTables(): Set wiiPtblLoaded = true.
I somehow forgot to do this, but it never caused problems before.
With the achievements code, it *does* cause problems, since the disc
image is closed before the achievements are checked.

WiiPartition: Changed EncKey to an `enum class`. We did this before
in WiiWAD for WadType. We can't really do it for DiscType in GameCube
because DiscType is a bitfield.