[librpbase] Achievements: Implemented save() and load().
save() is called if an achievement's status changes, e.g. the count is
incremented or a bit in the bitfield changes from 0 to 1.

load() is called if the achievements have not been loaded and unlock()
has been called. I'll need to add another function to actually retrieve
the achievements later to display them.

save() saves using a custom binary format. This format is normally
obfuscated in release builds, and stored in "ach.bin" in the config
directory. In debug builds, it's stored unobfuscated in "achd.bin".

The obfuscation isn't that complex, so cheating is possible, but don't
do it. Thanks. :)