[libromdata] SNES: Japanese titles might have Shift-JIS titles.
We're now only marking control codes as invalid in the title field.
All other characters are valid.

If the ROM destination is Japan, then Shift-JIS decoding will be
attempted. If it fails, cp1252 will be used.

cp1252 is used for all other regions. (Previously, it was Latin-1.)

This fixes detection of a bunch of Japanese Super Famicom ROMs.