Rescale SNES title screens from 256x224 -> 320x224 or 512x448 -> 640x448.
SNES uses non-square pixels, so we need to rescale the title screens
to show them properly. This is done using bilinear interpolation to
prevent weird artifacting, since it's a non-integer scaling.

RomData: New IMGPF flag IMGPF_RESCALE_256to320_512to640, which indicates
this scaling is needed. It only has any effect if the source image is
256px or 512px wide, and it doesn't affect the image height.

TCreateThumbnail: rescaleImgClass() now takes a ScalingMethod parameter
to indicate the scaling method to use. Default is Nearest.

[kde] Don't use Qt::KeepAspectRatio. The GTK+ and Win32 versions don't
do this, and we can't have it set because the SNES scaling inherently
changes the aspect ratio. The aspect ratio *is* preserved elsewhere in
cases where it's needed.

FIXME: On Windows 7, there seems to be an issue where if the image isn't
square, it originally shows up as if it's square, and *then* it's
shrunken down to its correct aspect ratio. May need to pad it to make
it square...