[libromdata] SufamiTurbo: New parser for Sufami Turbo mini-carts.
Sufami Turbo is an expansion cartridge for the Super Famicom, so I'm
considering this as part of SNES.

External title screens are stored in the "snes/title/ST/" directory
on the RPDB server.

Moved formatBankSizeKiB() from NES.cpp to TextFuncs.hpp and renamed
it to formatFileSizeKiB(). SufamiTurbo uses this function for the
ROM and SRAM size fields. "KiB" is now localized, too.

SNES: Minor comment fixes.

[libcachecommon] CacheKeys: Escape '#'. "Poi Poi Ninja World" for
Sufami Turbo has '#' in its internal title.