[libromdata] Console: Converted lots of enums to enum class.
Cast the "Unknown" enum value to int instead of using -1 directly.
I should go through all of the files later and check for any more
direct usage of -1 and change them if possible.

PlayStationSave: Reworked the Raw check to return Raw if it *is*
Raw instead of returning -1 if it's not.