RefBase.hpp: Added macros to make unref() simpler.
UNREF(): Check if the object is NULL; if it isn't, unref() it.
UNREF_AND_NULL(): Same, but set the variable to NULL afterwards.

UNREF_AND_NULL_NOCHK(): Same as UNREF_AND_NULL(), but skips the check.
Only use this if the object is known to be not NULL.

EXE: Added an extra UNREF_AND_NULL_NOCHK().

FileFormat subclasses: Don't bother setting isValid = false in cases where
the file isn't supported and we're not using an isRomSupported_static()
function, since the base class initializes isValid = false.