[libromdata] KeyStoreUI: Added more validation.
While the UI did validate for hex digits and kanji, it did *not* validate
for key length. Instead of adding that to the UI, add it here, and also
validate hex digits and kanji while we're at it.

TODO: Handle longer keys if/when they're implemented later.

This fixes issue #193: rp-config: keys longer than 16 bytes are validated