[gtk3] RpThunarProvider: Create a blank ThunarxPropertyPage instead of subclassing it.
It turns out we can use the same method for Thunarx that we're using
for Nautilus: Create a blank property page, then add a RomDataView
object to it.

This might not have the change notification from ThunarxFileInfo, but
I'm not sure if that even worked as I thought it did.

This significantly reduces the complexity of the gtk3 UI frontend.