[libromdata] GameCube: Added detection of RVZ files.
RVZ is an updated version of WIA designed for Dolphin. The header is
essentially the same as WIA, so we're using the same detection code.

Rearrange some code to move WIA and RVZ after the other checks.

Fixed the WIA disc type check. We should have been checking a BE32
value, not a single byte. This realistically didn't cause any problems,
since if this check failed (and it always did), it fell back to checking
for the GameCube and Wii magic numbers.